Add graphModule.js
Created makeGraph factory which receive the gameboard as argument.
Inside is a grid variable which stores the total grid a board has;
A maxRow and maxCol to store how many row and column the board has.
Created a Vertex factory which return the vertex and its vertex edges.
Created the getBoardValue function to return the value of the board per
grid and also be use to filter if the row and col value exeeds the board.
Created the buildGraph funtion which returns the Adjancency list.