Add knight.js
Created a Knight factory that will take in the board and will return a
move method.
It has a variable possibleMoves which stores the Adjacency List of how a
knight chess piece moves.
The method move receives start and end position of the chessboard and
will facilitate in calling appropriate functions to trace, get, and
print the moves needed from the start to reach the end position.
The traceMoves function carries out a Breadth first Search to trace the
required moves from start to end.
The getMoves recursive function, receives the vertex in which it will
put all the moves in the moveArray.
The PrintMoves function is to print the result in the console.
The clearVertexSrcDist function is for resetting the src and distance of
the vertices.